+++ CHANGELOG.md
@@ -3,7 +3,14 @@
 ### Added
 
 - Added automatic changelog & releases
+- Added NoOp module
+- Added VFS module
+- Added pipeline module
+- Added `ExperimentalSsioApi` annotation for incubating APIs
+- Added RandomAsyncSource
+- Added DiscardingAsyncSink
 
 ### Changed
 
-- Updated to Karma Conventions 1.11.0
+- Updated to Karma Conventions 1.11.4
+- Split IO APIs into separate module